pyproject.toml
[project]
name = "goit-pythonweb-hw-06"
version = "0.1.0"
description = ""
authors = [
    {name = "mazurkevych30",email = "132198562+mazurkevych30@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "faker (>=37.0.2,<38.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
